+++ config/releases.exs
@@ -5,7 +5,8 @@ config :td_dd, TdDd.Repo,
   username: System.fetch_env!("DB_USER"),
   password: System.fetch_env!("DB_PASSWORD"),
   database: System.fetch_env!("DB_NAME"),
-  hostname: System.fetch_env!("DB_HOST")
+  hostname: System.fetch_env!("DB_HOST"),
+  pool_size: System.get_env("DB_POOL_SIZE", "8") |> String.to_integer()
 
 config :td_dd, TdDd.Auth.Guardian, secret_key: System.fetch_env!("GUARDIAN_SECRET_KEY")
 
@@ -17,12 +18,5 @@ config :td_dd, :audit_service,
   audit_port: System.fetch_env!("API_AUDIT_PORT")
 
 config :td_cache, :event_stream, consumer_id: System.fetch_env!("HOSTNAME")
-
-config :bolt_sips, Bolt,
-  hostname: System.get_env("NEO4J_HOST"),
-  basic_auth: [
-    username: System.get_env("NEO4J_USER"),
-    password: System.get_env("NEO4J_PASSWORD")
-  ]
 
 config :td_dd, import_dir: System.get_env("IMPORT_DIR")