+++ lib/td_dd/lineage/graph_data/nodes.ex
@@ -39,8 +39,8 @@ defmodule TdDd.Lineage.GraphData.Nodes do
   defp children(%Graph{} = t, id, roots) do
     t
     |> do_children(id, roots)
-    |> Enum.filter(&visible?(t, &1))
-    |> Enum.map(&Graph.vertex(t, &1, :data))
+    |> Enum.reject(&hidden?(t, &1))
+    |> Enum.map(&Graph.vertex(t, &1))
     |> Enum.map(&node_label/1)
     |> Enum.sort_by(&sortable/1)
     |> Enum.group_by(&class/1, &Map.delete(&1, :class))
@@ -49,23 +49,17 @@ defmodule TdDd.Lineage.GraphData.Nodes do
   defp do_children(%Graph{} = _t, nil, roots), do: roots
   defp do_children(%Graph{} = t, id, _roots), do: Graph.out_neighbours(t, id)
 
-  defp visible?(%Graph{} = t, id) do
-    case Graph.vertex(t, id, :data) do
-      %{properties: %{"select_hidden" => true}} -> false
-      _ -> true
-    end
+  defp hidden?(%Graph{} = t, id) do
+    Graph.vertex(t, id, "hidden") == true
   end
 
   defp class(%{class: "Group"}), do: :groups
   defp class(%{class: "Resource"}), do: :resources
 
-  defp node_label(%{data: data}), do: node_label(data)
-
-  defp node_label(%{properties: properties, labels: [class], id: id}) do
-    properties
-    |> Map.take(["name", "external_id", "type"])
-    |> Map.put(:class, class)
-    |> Map.put(:id, id)
+  defp node_label(%{label: label, id: id}) do
+    label
+    |> Map.take(["name", "type", :class])
+    |> Map.put("external_id", id)
   end
 
   defp node_label(_), do: %{}