+++ lib/td_dd_web/controllers/fallback_controller.ex
@@ -6,6 +6,8 @@ defmodule TdDdWeb.FallbackController do
   """
   use TdDdWeb, :controller
 
+  require Logger
+
   def call(conn, {:error, %Ecto.Changeset{} = changeset}) do
     conn
     |> put_status(:unprocessable_entity)
@@ -15,5 +17,14 @@ defmodule TdDdWeb.FallbackController do
 
   def call(conn, {:error, :not_found}) do
     render_error(conn, :not_found)
+  end
+
+  def call(conn, {:can, :false}) do
+    render_error(conn, :forbidden)
+  end
+
+  def call(conn, {:cp, {:error, error}}) do
+    Logger.warn("File copy operation failed with error #{inspect(error)}")
+    render_error(conn, :insufficient_storage)
   end
 end